+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - you get a cake for your birthday (@growling_grizzly)
 - you can now /marry someone (@growling_grizzly)
 - many new features in /tmdb (@growling_grizzly)
+- guild member slots increase requirements for level up
 
 ## other